Anubisafe/_alpha_ANUBIS/Alpha/test/requirements.txt
cryptography==2.1.4
h5py==2.9.0
intel.tracecli==1.1919.135
ipaddress==1.0.17
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.0
numpy==1.16.5
olefile==0.45.1
pexpect==4.7.0
Pillow==5.1.0
protobuf==3.9.1
ptyprocess==0.6.0
py-dmidecode==0.0.2
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
pyOpenSSL==17.5.0
pytz==2018.3
SecretStorage==2.3.1
six==1.12.0
Sphinx==1.6.7
starkbank-ecdsa==0.1.6
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
